+++ examples/07_router/routes/bar.tsx
@@ -2,12 +2,12 @@ import { ReactNode } from "react";
 
 import { Counter } from "../src/Counter.js";
 
-export default ({ children }: { children: ReactNode }) => {
-  return (
-    <div>
-      <h2>Bar</h2>
-      <Counter />
-      {children}
-    </div>
-  );
-};
+const Bar = ({ children }: { children: ReactNode }) => (
+  <div>
+    <h2>Bar</h2>
+    <Counter />
+    {children}
+  </div>
+);
+
+export default Bar;