+++ postcss.config.js
@@ -2,5 +2,5 @@ export default {
   plugins: {
     tailwindcss: {},
     autoprefixer: {},
-  }
-}
+  },
+};