+++ src/types.d.ts
@@ -3,7 +3,8 @@ declare module "react-server-dom-webpack/server";
 declare module "react-server-dom-webpack/server.node.unbundled";
 declare module "react-server-dom-webpack/client";
 
-// FIXME is this too naive?
+// FIXME Is this too naive? How can we avoid the eslint warning?
+// eslint-disable-next-line @typescript-eslint/no-unused-vars
 namespace JSX {
   export type ElementType =
     | string